+++ test/api/test_thread.cpp
@@ -1,0 +1,151 @@
+// SPDX-FileCopyrightText: 2024 Arm Limited and/or its affiliates <open-source-office@arm.com>
+//
+// SPDX-License-Identifier: Apache-2.0
+
+#include <gmock/gmock.h>
+#include <gtest/gtest.h>
+
+#include <thread>
+
+#include "framework/array.h"
+#include "framework/generator.h"
+#include "kleidicv/kleidicv.h"
+#include "kleidicv_thread/kleidicv_thread.h"
+#include "multithreading_fake.h"
+
+// Tuple of width, height, thread count.
+typedef std::tuple<unsigned, unsigned, unsigned> P;
+
+class Thread : public testing::TestWithParam<P> {
+ public:
+  template <typename SrcT, typename DstT, typename SingleThreadedFunc,
+            typename MultithreadedFunc, typename... Args>
+  void check_unary_op(SingleThreadedFunc single_threaded_func,
+                      MultithreadedFunc multithreaded_func, size_t src_channels,
+                      size_t dst_channels, Args... args) {
+    unsigned width = 0, height = 0, thread_count = 0;
+    std::tie(width, height, thread_count) = GetParam();
+    test::Array2D<SrcT> src(size_t{width} * src_channels, height);
+    test::Array2D<DstT> dst_single(size_t{width} * dst_channels, height),
+        dst_multi(size_t{width} * dst_channels, height);
+
+    test::PseudoRandomNumberGenerator<SrcT> generator;
+    src.fill(generator);
+
+    kleidicv_error_t single_result =
+        single_threaded_func(src.data(), src.stride(), dst_single.data(),
+                             dst_single.stride(), width, height, args...);
+
+    kleidicv_error_t multi_result = multithreaded_func(
+        src.data(), src.stride(), dst_multi.data(), dst_multi.stride(), width,
+        height, args..., get_multithreading_fake(thread_count));
+
+    EXPECT_EQ(KLEIDICV_OK, single_result);
+    EXPECT_EQ(KLEIDICV_OK, multi_result);
+    EXPECT_EQ_ARRAY2D(dst_multi, dst_single);
+  }
+
+  template <typename SrcT, typename DstT, typename SingleThreadedFunc,
+            typename MultithreadedFunc, typename... Args>
+  void check_binary_op(SingleThreadedFunc single_threaded_func,
+                       MultithreadedFunc multithreaded_func,
+                       size_t src_channels, size_t dst_channels, Args... args) {
+    unsigned width = 0, height = 0, thread_count = 0;
+    std::tie(width, height, thread_count) = GetParam();
+    test::Array2D<SrcT> src_a(size_t{width} * src_channels, height),
+        src_b(size_t{width} * src_channels, height);
+    test::Array2D<DstT> dst_single(size_t{width} * dst_channels, height),
+        dst_multi(size_t{width} * dst_channels, height);
+
+    test::PseudoRandomNumberGenerator<SrcT> generator;
+    src_a.fill(generator);
+    src_b.fill(generator);
+
+    kleidicv_error_t single_result = single_threaded_func(
+        src_a.data(), src_a.stride(), src_b.data(), src_b.stride(),
+        dst_single.data(), dst_single.stride(), width, height, args...);
+
+    kleidicv_error_t multi_result = multithreaded_func(
+        src_a.data(), src_a.stride(), src_b.data(), src_b.stride(),
+        dst_multi.data(), dst_multi.stride(), width, height, args...,
+        get_multithreading_fake(thread_count));
+
+    EXPECT_EQ(KLEIDICV_OK, single_result);
+    EXPECT_EQ(KLEIDICV_OK, multi_result);
+    EXPECT_EQ_ARRAY2D(dst_multi, dst_single);
+  }
+};
+
+#define TEST_UNARY_OP(suffix, SrcT, DstT, ...)                              \
+  TEST_P(Thread, suffix) {                                                  \
+    check_unary_op<SrcT, DstT>(kleidicv_##suffix, kleidicv_thread_##suffix, \
+                               __VA_ARGS__);                                \
+  }
+
+#define TEST_BINARY_OP(suffix, T, ...)                                 \
+  TEST_P(Thread, suffix) {                                             \
+    check_binary_op<T, T>(kleidicv_##suffix, kleidicv_thread_##suffix, \
+                          __VA_ARGS__);                                \
+  }
+
+TEST_UNARY_OP(gray_to_rgb_u8, uint8_t, uint8_t, 1, 3);
+TEST_UNARY_OP(gray_to_rgba_u8, uint8_t, uint8_t, 1, 4);
+TEST_UNARY_OP(rgb_to_bgr_u8, uint8_t, uint8_t, 3, 3);
+TEST_UNARY_OP(rgb_to_rgb_u8, uint8_t, uint8_t, 3, 3);
+TEST_UNARY_OP(rgba_to_bgra_u8, uint8_t, uint8_t, 4, 4);
+TEST_UNARY_OP(rgba_to_rgba_u8, uint8_t, uint8_t, 4, 4);
+TEST_UNARY_OP(rgb_to_bgra_u8, uint8_t, uint8_t, 3, 4);
+TEST_UNARY_OP(rgb_to_rgba_u8, uint8_t, uint8_t, 3, 4);
+TEST_UNARY_OP(rgba_to_bgr_u8, uint8_t, uint8_t, 4, 3);
+TEST_UNARY_OP(rgba_to_rgb_u8, uint8_t, uint8_t, 4, 3);
+TEST_UNARY_OP(yuv_to_bgr_u8, uint8_t, uint8_t, 3, 3);
+TEST_UNARY_OP(yuv_to_rgb_u8, uint8_t, uint8_t, 3, 3);
+TEST_UNARY_OP(bgr_to_yuv_u8, uint8_t, uint8_t, 3, 3);
+TEST_UNARY_OP(rgb_to_yuv_u8, uint8_t, uint8_t, 3, 3);
+TEST_UNARY_OP(bgra_to_yuv_u8, uint8_t, uint8_t, 4, 3);
+TEST_UNARY_OP(rgba_to_yuv_u8, uint8_t, uint8_t, 4, 3);
+TEST_UNARY_OP(threshold_binary_u8, uint8_t, uint8_t, 1, 1, 100, 200);
+TEST_UNARY_OP(scale_u8, uint8_t, uint8_t, 1, 1, 0.5F, 3.5F);
+TEST_UNARY_OP(scale_f32, float, float, 1, 1, 0.123F, 45.6789F);
+TEST_UNARY_OP(exp_f32, float, float, 1, 1);
+TEST_UNARY_OP(float_conversion_f32_s8, float, int8_t, 1, 1);
+TEST_UNARY_OP(float_conversion_f32_u8, float, uint8_t, 1, 1);
+TEST_UNARY_OP(float_conversion_s8_f32, int8_t, float, 1, 1);
+TEST_UNARY_OP(float_conversion_u8_f32, uint8_t, float, 1, 1);
+
+TEST_BINARY_OP(saturating_add_s8, int8_t, 1, 1);
+TEST_BINARY_OP(saturating_add_u8, uint8_t, 1, 1);
+TEST_BINARY_OP(saturating_add_s16, int16_t, 1, 1);
+TEST_BINARY_OP(saturating_add_u16, uint16_t, 1, 1);
+TEST_BINARY_OP(saturating_add_s32, int32_t, 1, 1);
+TEST_BINARY_OP(saturating_add_u32, uint32_t, 1, 1);
+TEST_BINARY_OP(saturating_add_s64, int64_t, 1, 1);
+TEST_BINARY_OP(saturating_add_u64, uint64_t, 1, 1);
+TEST_BINARY_OP(saturating_sub_s8, int8_t, 1, 1);
+TEST_BINARY_OP(saturating_sub_u8, uint8_t, 1, 1);
+TEST_BINARY_OP(saturating_sub_s16, int16_t, 1, 1);
+TEST_BINARY_OP(saturating_sub_u16, uint16_t, 1, 1);
+TEST_BINARY_OP(saturating_sub_s32, int32_t, 1, 1);
+TEST_BINARY_OP(saturating_sub_u32, uint32_t, 1, 1);
+TEST_BINARY_OP(saturating_sub_s64, int64_t, 1, 1);
+TEST_BINARY_OP(saturating_sub_u64, uint64_t, 1, 1);
+TEST_BINARY_OP(saturating_absdiff_u8, uint8_t, 1, 1);
+TEST_BINARY_OP(saturating_absdiff_s8, int8_t, 1, 1);
+TEST_BINARY_OP(saturating_absdiff_u16, uint16_t, 1, 1);
+TEST_BINARY_OP(saturating_absdiff_s16, int16_t, 1, 1);
+TEST_BINARY_OP(saturating_absdiff_s32, int32_t, 1, 1);
+TEST_BINARY_OP(saturating_multiply_u8, uint8_t, 1, 1, 1.23);
+TEST_BINARY_OP(saturating_multiply_s8, int8_t, 1, 1, -2.34);
+TEST_BINARY_OP(saturating_multiply_u16, uint16_t, 1, 1, 0.321);
+TEST_BINARY_OP(saturating_multiply_s16, int16_t, 1, 1, -0.543);
+TEST_BINARY_OP(saturating_multiply_s32, int32_t, 1, 1, -0.0123);
+TEST_BINARY_OP(bitwise_and, uint8_t, 1, 1);
+TEST_BINARY_OP(compare_equal_u8, uint8_t, 1, 1);
+TEST_BINARY_OP(compare_greater_u8, uint8_t, 1, 1);
+TEST_BINARY_OP(saturating_add_abs_with_threshold_s16, int16_t, 1, 1, 123);
+
+INSTANTIATE_TEST_SUITE_P(, Thread,
+                         testing::Values(P{1, 1, 1}, P{1, 2, 1}, P{1, 2, 2},
+                                         P{2, 1, 2}, P{2, 2, 1}, P{1, 3, 2},
+                                         P{2, 3, 1}, P{6, 4, 1}, P{4, 5, 2},
+                                         P{2, 6, 3}, P{1, 7, 4}, P{12, 34, 5}));